pyproject.toml
[project]
name = "effib0-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface-hub>=0.34.4",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
